Pod listing!! It works!
Had to add a clientset and a podOpts, then use the clientset to get
the corev1 api contents.

Then I got permissions errors from the pod deployment; it was trying
to reach the pods endpoint but could not. So I added the rbac.yaml
file with a service account that has a role and a rolebinding, then
I updated the pod to use the new service account.

Funnily, using rawyaml it looks like skaffold applies the files in
order, so I had to reorder the rawyaml.